Extracted factory to interface. Beefed up exception handling
