+++ src/main/java/skaro/frenbot/commands/Command.java
@@ -9,6 +9,5 @@ public interface Command {
 
 	Mono<Message> execute();
 	void setArguments(List<String> arguments);
-	String getName();
 	
 }
