+++ src/main/java/skaro/frenbot/commands/parsers/ParserException.java
@@ -4,12 +4,12 @@ public class ParserException extends RuntimeException {
 
 	private static final long serialVersionUID = 8887248364227285285L;
 	
-	public ParserException(String arguments) {
-		super("Could not parse the string: "+ arguments);
+	public ParserException(String arguments, Exception e) {
+		super("Exception "+ e.getClass().getSimpleName() +" - Could not parse the string: "+ arguments);
 	}
 	
-	public ParserException(String[] arguments) {
-		super("Could not parse the string: "+ String.join(" ", arguments));
+	public ParserException(String[] arguments, Exception e) {
+		super("Exception "+ e.getClass().getSimpleName() +" - Could not parse the string: "+ String.join(" ", arguments));
 	}
 
 }
